refactor: [M3-5945] - Parameterize DBaaS tests for multiple engines (#8900)
* M3-5945: Parameterize DBaaS tests for multiple engines
* Changed based on comments
