+++ packages/manager/cypress/e2e/databases/create-database.spec.ts
@@ -1,65 +1,12 @@
-import { ClusterSize, Engine } from '@linode/api-v4/types';
 import { databaseInstanceFactory } from 'src/factories/databases';
 import { eventFactory } from 'src/factories/events';
-import { randomLabel } from 'support/util/random';
 import { sequentialStub } from 'support/stubs/sequential-stub';
 import { paginateResponse } from 'support/util/paginate';
 import { apiMatcher } from 'support/util/intercepts';
-
-interface databaseClusterConfiguration {
-  label: string;
-  linodeType: string;
-  clusterSize: ClusterSize;
-  dbType: Engine;
-  region: string;
-  regionTypeahead: string;
-  engine: string;
-  version: string;
-}
-
-// Array of database cluster configurations for which to test creation.
-const databaseConfigurations: databaseClusterConfiguration[] = [
-  {
-    label: randomLabel(),
-    linodeType: 'g6-nanode-1',
-    clusterSize: 1,
-    dbType: 'mysql',
-    region: 'us-east',
-    regionTypeahead: 'Newark',
-    engine: 'MySQL',
-    version: '8',
-  },
-  {
-    label: randomLabel(),
-    linodeType: 'g6-dedicated-16',
-    clusterSize: 3,
-    dbType: 'mysql',
-    region: 'us-southeast',
-    regionTypeahead: 'Atlanta',
-    engine: 'MySQL',
-    version: '5',
-  },
-  // {
-  //   label: randomLabel(),
-  //   linodeType: 'g6-dedicated-16',
-  //   clusterSize: 1,
-  //   dbType: 'mongodb',
-  //   regionTypeahead: 'Atlanta',
-  //   region: 'us-southeast',
-  //   engine: 'MongoDB',
-  //   version: '4',
-  // },
-  {
-    label: randomLabel(),
-    linodeType: 'g6-nanode-1',
-    clusterSize: 3,
-    dbType: 'postgresql',
-    region: 'us-east',
-    regionTypeahead: 'Newark',
-    engine: 'PostgreSQL',
-    version: '13',
-  },
-];
+import {
+  databaseClusterConfiguration,
+  databaseConfigurations,
+} from 'support/constants/databases';
 
 describe('create a database cluster, mocked data', () => {
   databaseConfigurations.forEach(
