+++ packages/manager/cypress/e2e/databases/update-database.spec.ts
@@ -2,10 +2,9 @@
  * @file DBaaS integration tests for update operations.
  */
 
-import { regions, regionsMap } from 'support/constants/regions';
+import { regionsMap } from 'support/constants/regions';
 import {
   randomLabel,
-  randomItem,
   randomNumber,
   randomIp,
   randomString,
@@ -20,6 +19,10 @@ import {
   mockUpdateDatabase,
   mockUpdateProvisioningDatabase,
 } from 'support/intercepts/databases';
+import {
+  databaseClusterConfiguration,
+  databaseConfigurations,
+} from 'support/constants/databases';
 
 // @TODO Consider moving this to 'support/constants' or similar.
 // const databaseTypes = ['mysql', 'postgresql', 'mongodb'];
@@ -140,223 +143,234 @@ const resetRootPassword = () => {
 };
 
 describe('Update database clusters', () => {
-  /*
-   * - Tests active database update UI flows using mocked data.
-   * - Confirms that users can change database label.
-   * - Confirms that users can change access controls.
-   * - Confirms that users can reset root passwords.
-   * - Confirms that users can change maintenance schedules.
-   */
-  it('Can update active database clusters', () => {
-    const initialLabel = randomLabel();
-    const updatedLabel = randomLabel();
-    const allowedIp = randomIp();
-    const newAllowedIp = randomIp();
-    const engineType = randomItem(databaseTypes);
-    const initialPassword = randomString(16);
-    const database = databaseFactory.build({
-      id: randomNumber(1, 1000),
-      type: 'g6-nanode-1',
-      label: initialLabel,
-      region: randomItem(regions),
-      engine: engineType,
-      status: 'active',
-      allow_list: [allowedIp],
-    });
+  databaseConfigurations.forEach(
+    (configuration: databaseClusterConfiguration) => {
+      describe(`updates a ${configuration.linodeType} ${configuration.engine} v${configuration.version}.x ${configuration.clusterSize}-node cluster at ${configuration.region}`, () => {
+        /*
+         * - Tests active database update UI flows using mocked data.
+         * - Confirms that users can change database label.
+         * - Confirms that users can change access controls.
+         * - Confirms that users can reset root passwords.
+         * - Confirms that users can change maintenance schedules.
+         */
+        it('Can update active database clusters', () => {
+          const initialLabel = configuration.label;
+          const updatedLabel = randomLabel();
+          const allowedIp = randomIp();
+          const newAllowedIp = randomIp();
+          const initialPassword = randomString(16);
+          const database = databaseFactory.build({
+            id: randomNumber(1, 1000),
+            type: configuration.linodeType,
+            label: initialLabel,
+            region: configuration.region,
+            engine: configuration.dbType,
+            status: 'active',
+            allow_list: [allowedIp],
+          });
 
-    mockGetDatabase(database).as('getDatabase');
-    mockResetPassword(database.id, database.engine).as('resetRootPassword');
-    mockGetDatabaseCredentials(
-      database.id,
-      database.engine,
-      initialPassword
-    ).as('getCredentials');
+          mockGetDatabase(database).as('getDatabase');
+          mockResetPassword(database.id, database.engine).as(
+            'resetRootPassword'
+          );
+          mockGetDatabaseCredentials(
+            database.id,
+            database.engine,
+            initialPassword
+          ).as('getCredentials');
 
-    cy.visitWithLogin(`/databases/${database.engine}/${database.id}`);
-    cy.wait('@getDatabase');
+          cy.visitWithLogin(`/databases/${database.engine}/${database.id}`);
+          cy.wait('@getDatabase');
 
-    cy.get('[data-qa-cluster-config]').within(() => {
-      cy.findByText(regionsMap[database.region]).should('be.visible');
-    });
+          cy.get('[data-qa-cluster-config]').within(() => {
+            cy.findByText(regionsMap[database.region]).should('be.visible');
+          });
 
-    cy.get('[data-qa-connection-details]').within(() => {
-      // "Show" button should be enabled to reveal password when DB is active.
-      cy.findByText('Show')
-        .closest('button')
-        .should('be.visible')
-        .should('be.enabled')
-        .click();
+          cy.get('[data-qa-connection-details]').within(() => {
+            // "Show" button should be enabled to reveal password when DB is active.
+            cy.findByText('Show')
+              .closest('button')
+              .should('be.visible')
+              .should('be.enabled')
+              .click();
 
-      cy.wait('@getCredentials');
-      cy.findByText(`= ${initialPassword}`);
-    });
+            cy.wait('@getCredentials');
+            cy.findByText(`= ${initialPassword}`);
+          });
 
-    mockUpdateDatabase(database.id, database.engine, {
-      ...database,
-      label: updatedLabel,
-    }).as('updateDatabaseLabel');
-    updateDatabaseLabel(initialLabel, updatedLabel);
-    cy.wait('@updateDatabaseLabel');
-    cy.get('[data-qa-header]')
-      .should('be.visible')
-      .should('have.text', updatedLabel);
+          mockUpdateDatabase(database.id, database.engine, {
+            ...database,
+            label: updatedLabel,
+          }).as('updateDatabaseLabel');
+          updateDatabaseLabel(initialLabel, updatedLabel);
+          cy.wait('@updateDatabaseLabel');
+          cy.get('[data-qa-header]')
+            .should('be.visible')
+            .should('have.text', updatedLabel);
 
-    // Remove allowed IP, manage IP access control.
-    mockUpdateDatabase(database.id, database.engine, {
-      ...database,
-      allow_list: [],
-    }).as('updateDatabaseAllowedIp');
-    removeAllowedIp(allowedIp);
-    cy.wait('@updateDatabaseAllowedIp');
+          // Remove allowed IP, manage IP access control.
+          mockUpdateDatabase(database.id, database.engine, {
+            ...database,
+            allow_list: [],
+          }).as('updateDatabaseAllowedIp');
+          removeAllowedIp(allowedIp);
+          cy.wait('@updateDatabaseAllowedIp');
 
-    mockUpdateDatabase(database.id, database.engine, {
-      ...database,
-      allow_list: [newAllowedIp],
-    }).as('updateAccessControl');
-    manageAccessControl([newAllowedIp]);
-    cy.wait('@updateAccessControl');
-    cy.get('[data-qa-access-controls]').within(() => {
-      cy.findByText(newAllowedIp).should('be.visible');
-    });
+          mockUpdateDatabase(database.id, database.engine, {
+            ...database,
+            allow_list: [newAllowedIp],
+          }).as('updateAccessControl');
+          manageAccessControl([newAllowedIp]);
+          cy.wait('@updateAccessControl');
+          cy.get('[data-qa-access-controls]').within(() => {
+            cy.findByText(newAllowedIp).should('be.visible');
+          });
 
-    // Navigate to "Settings" tab.
-    ui.tabList.findTabByTitle('Settings').click();
+          // Navigate to "Settings" tab.
+          ui.tabList.findTabByTitle('Settings').click();
 
-    // Reset root password.
-    resetRootPassword();
-    cy.wait('@resetRootPassword');
+          // Reset root password.
+          resetRootPassword();
+          cy.wait('@resetRootPassword');
 
-    // Change maintenance.
-    mockUpdateDatabase(database.id, database.engine, database).as(
-      'updateDatabaseMaintenance'
-    );
-    cy.findByText('Monthly').should('be.visible').click();
+          // Change maintenance.
+          mockUpdateDatabase(database.id, database.engine, database).as(
+            'updateDatabaseMaintenance'
+          );
+          cy.findByText('Monthly').should('be.visible').click();
 
-    ui.button
-      .findByTitle('Save Changes')
-      .should('be.visible')
-      .should('be.enabled')
-      .click();
+          ui.button
+            .findByTitle('Save Changes')
+            .should('be.visible')
+            .should('be.enabled')
+            .click();
 
-    cy.wait('@updateDatabaseMaintenance');
-    ui.toast.assertMessage('Maintenance Window settings saved successfully.');
-  });
+          cy.wait('@updateDatabaseMaintenance');
+          ui.toast.assertMessage(
+            'Maintenance Window settings saved successfully.'
+          );
+        });
 
-  /*
-   * - Tests provisioning database update UI flows using mocked data.
-   * - Confirms that database update flows work under error conditions.
-   * - Confirms that users cannot change database label for provisioning DBs.
-   * - Confirms that users cannot change access controls for provisioning DBs.
-   * - Confirms that users cannot reset root passwords for provisioning DBs.
-   * - Confirms that users cannot change maintenance schedules for provisioning DBs.
-   */
-  it('Cannot update database clusters while they are provisioning', () => {
-    const initialLabel = randomLabel();
-    const updateAttemptLabel = randomLabel();
-    const allowedIp = randomIp();
-    const engineType = randomItem(databaseTypes);
-    const database = databaseFactory.build({
-      id: randomNumber(1, 1000),
-      type: 'g6-nanode-1',
-      label: initialLabel,
-      region: randomItem(regions),
-      engine: engineType,
-      status: 'provisioning',
-      allow_list: [allowedIp],
-      hosts: {
-        primary: undefined,
-        secondary: undefined,
-      },
-    });
+        /*
+         * - Tests provisioning database update UI flows using mocked data.
+         * - Confirms that database update flows work under error conditions.
+         * - Confirms that users cannot change database label for provisioning DBs.
+         * - Confirms that users cannot change access controls for provisioning DBs.
+         * - Confirms that users cannot reset root passwords for provisioning DBs.
+         * - Confirms that users cannot change maintenance schedules for provisioning DBs.
+         */
+        it('Cannot update database clusters while they are provisioning', () => {
+          const initialLabel = configuration.label;
+          const updateAttemptLabel = randomLabel();
+          const allowedIp = randomIp();
+          const database = databaseFactory.build({
+            id: randomNumber(1, 1000),
+            type: configuration.linodeType,
+            label: initialLabel,
+            region: configuration.region,
+            engine: configuration.dbType,
+            status: 'provisioning',
+            allow_list: [allowedIp],
+            hosts: {
+              primary: undefined,
+              secondary: undefined,
+            },
+          });
 
-    const errorMessage =
-      'Your database is provisioning; please wait until provisioning is complete to perform this operation.';
-    const hostnameRegex = /your hostnames? will appear here once (it is|they are) available./i;
+          const errorMessage =
+            'Your database is provisioning; please wait until provisioning is complete to perform this operation.';
+          const hostnameRegex = /your hostnames? will appear here once (it is|they are) available./i;
 
-    mockGetDatabase(database).as('getDatabase');
+          mockGetDatabase(database).as('getDatabase');
 
-    mockUpdateProvisioningDatabase(
-      database.id,
-      database.engine,
-      errorMessage
-    ).as('updateDatabase');
+          mockUpdateProvisioningDatabase(
+            database.id,
+            database.engine,
+            errorMessage
+          ).as('updateDatabase');
 
-    mockResetPasswordProvisioningDatabase(
-      database.id,
-      database.engine,
-      errorMessage
-    ).as('resetRootPassword');
+          mockResetPasswordProvisioningDatabase(
+            database.id,
+            database.engine,
+            errorMessage
+          ).as('resetRootPassword');
 
-    cy.visitWithLogin(`/databases/${database.engine}/${database.id}`);
-    cy.wait('@getDatabase');
+          cy.visitWithLogin(`/databases/${database.engine}/${database.id}`);
+          cy.wait('@getDatabase');
 
-    // Cannot update database label.
-    updateDatabaseLabel(initialLabel, updateAttemptLabel);
-    cy.wait('@updateDatabase');
-    cy.findByText(errorMessage).should('be.visible');
-    cy.get('[data-qa-cancel-edit="true"]')
-      .should('be.visible')
-      .should('be.enabled')
-      .click();
+          // Cannot update database label.
+          updateDatabaseLabel(initialLabel, updateAttemptLabel);
+          cy.wait('@updateDatabase');
+          cy.findByText(errorMessage).should('be.visible');
+          cy.get('[data-qa-cancel-edit="true"]')
+            .should('be.visible')
+            .should('be.enabled')
+            .click();
 
-    cy.get('[data-qa-connection-details]').within(() => {
-      // DBaaS hostnames are not available until database/cluster has provisioned.
-      cy.findByText(hostnameRegex).should('be.visible');
+          cy.get('[data-qa-connection-details]').within(() => {
+            // DBaaS hostnames are not available until database/cluster has provisioned.
+            cy.findByText(hostnameRegex).should('be.visible');
 
-      // DBaaS passwords cannot be revealed until database/cluster has provisioned.
-      cy.findByText('Show')
-        .closest('button')
-        .should('be.visible')
-        .should('be.disabled');
-    });
+            // DBaaS passwords cannot be revealed until database/cluster has provisioned.
+            cy.findByText('Show')
+              .closest('button')
+              .should('be.visible')
+              .should('be.disabled');
+          });
 
-    // Cannot add or remove allowed IPs before database/cluster has provisioned.
-    removeAllowedIp(allowedIp);
-    cy.wait('@updateDatabase');
-    ui.dialog
-      .findByTitle(`Remove IP Address ${allowedIp}`)
-      .should('be.visible')
-      .within(() => {
-        cy.findByText(errorMessage).should('be.visible');
-        ui.buttonGroup.findButtonByTitle('Cancel').should('be.visible').click();
+          // Cannot add or remove allowed IPs before database/cluster has provisioned.
+          removeAllowedIp(allowedIp);
+          cy.wait('@updateDatabase');
+          ui.dialog
+            .findByTitle(`Remove IP Address ${allowedIp}`)
+            .should('be.visible')
+            .within(() => {
+              cy.findByText(errorMessage).should('be.visible');
+              ui.buttonGroup
+                .findButtonByTitle('Cancel')
+                .should('be.visible')
+                .click();
+            });
+
+          manageAccessControl([randomIp()], 1);
+          cy.wait('@updateDatabase');
+          ui.drawer.findByTitle('Manage Access Controls').within(() => {
+            cy.findByText(errorMessage).should('be.visible');
+            ui.drawerCloseButton.find().click();
+          });
+
+          // Navigate to "Settings" tab.
+          ui.tabList.findTabByTitle('Settings').click();
+
+          // Cannot reset root password before database/cluster has provisioned.
+          resetRootPassword();
+          cy.wait('@resetRootPassword');
+          ui.dialog
+            .findByTitle('Reset Root Password')
+            .should('be.visible')
+            .within(() => {
+              cy.findByText(errorMessage).should('be.visible');
+
+              ui.buttonGroup
+                .findButtonByTitle('Cancel')
+                .should('be.visible')
+                .should('be.enabled')
+                .click();
+            });
+
+          // Cannot change maintenance schedule before database/cluster has provisioned.
+          cy.findByText('Monthly').should('be.visible').click();
+
+          ui.button
+            .findByTitle('Save Changes')
+            .should('be.visible')
+            .should('be.enabled')
+            .click();
+
+          cy.wait('@updateDatabase');
+          cy.findByText(errorMessage).should('be.visible');
+        });
       });
-
-    manageAccessControl([randomIp()], 1);
-    cy.wait('@updateDatabase');
-    ui.drawer.findByTitle('Manage Access Controls').within(() => {
-      cy.findByText(errorMessage).should('be.visible');
-      ui.drawerCloseButton.find().click();
-    });
-
-    // Navigate to "Settings" tab.
-    ui.tabList.findTabByTitle('Settings').click();
-
-    // Cannot reset root password before database/cluster has provisioned.
-    resetRootPassword();
-    cy.wait('@resetRootPassword');
-    ui.dialog
-      .findByTitle('Reset Root Password')
-      .should('be.visible')
-      .within(() => {
-        cy.findByText(errorMessage).should('be.visible');
-
-        ui.buttonGroup
-          .findButtonByTitle('Cancel')
-          .should('be.visible')
-          .should('be.enabled')
-          .click();
-      });
-
-    // Cannot change maintenance schedule before database/cluster has provisioned.
-    cy.findByText('Monthly').should('be.visible').click();
-
-    ui.button
-      .findByTitle('Save Changes')
-      .should('be.visible')
-      .should('be.enabled')
-      .click();
-
-    cy.wait('@updateDatabase');
-    cy.findByText(errorMessage).should('be.visible');
-  });
+    }
+  );
 });
